+++ Texture/components/com_render1.ts
@@ -1,0 +1,1 @@
+../../core/components/com_render1.ts
