+++ Texture/game.ts
@@ -1,11 +1,11 @@
 import {GL_CULL_FACE, GL_DEPTH_TEST} from "../common/webgl.js";
+import {mat1_textured} from "../materials/mat1_textured.js";
 import {mesh_kulka} from "../meshes/kulka.js";
 import {Camera} from "./components/com_camera.js";
 import {loop_start, loop_stop} from "./core.js";
-import {mat1_textured} from "./materials/mat1_textured.js";
 import {sys_camera} from "./systems/sys_camera.js";
 import {sys_framerate} from "./systems/sys_framerate.js";
-import {sys_render} from "./systems/sys_render.js";
+import {sys_render} from "./systems/sys_render1.js";
 import {sys_rotate} from "./systems/sys_rotate.js";
 import {sys_transform} from "./systems/sys_transform.js";
 import {World} from "./world.js";
@@ -30,6 +30,9 @@ export class Game {
     Textures: Record<string, WebGLTexture> = {};
 
     Camera?: Camera;
+    // The rendering pipeline supports 8 lights.
+    LightPositions = new Float32Array(4 * 8);
+    LightDetails = new Float32Array(4 * 8);
 
     constructor() {
         document.addEventListener("visibilitychange", () =>
