--- Texture/systems/sys_render.ts
@@ -1,71 +1,0 @@
-import {Material} from "../../common/material.js";
-import {
-    GL_COLOR_BUFFER_BIT,
-    GL_DEPTH_BUFFER_BIT,
-    GL_TEXTURE0,
-    GL_TEXTURE_2D,
-    GL_UNSIGNED_SHORT,
-} from "../../common/webgl.js";
-import {RenderKind} from "../components/com_render.js";
-import {RenderTextured} from "../components/com_render_textured.js";
-import {Transform} from "../components/com_transform.js";
-import {Game} from "../game.js";
-import {TexturedLayout} from "../materials/layout_textured.js";
-import {Has} from "../world.js";
-
-const QUERY = Has.Transform | Has.Render;
-
-export function sys_render(game: Game, delta: number) {
-    game.Gl.clear(GL_COLOR_BUFFER_BIT | GL_DEPTH_BUFFER_BIT);
-    if (game.ViewportResized) {
-        game.Gl.viewport(0, 0, game.ViewportWidth, game.ViewportHeight);
-    }
-
-    // Keep track of the current material to minimize switching.
-    let current_material = null;
-    let current_front_face = null;
-
-    for (let i = 0; i < game.World.Signature.length; i++) {
-        if ((game.World.Signature[i] & QUERY) === QUERY) {
-            let transform = game.World.Transform[i];
-            let render = game.World.Render[i];
-
-            if (render.Material !== current_material) {
-                current_material = render.Material;
-                switch (render.Kind) {
-                    case RenderKind.Textured:
-                        use_textured(game, render.Material);
-                        break;
-                }
-            }
-
-            if (render.FrontFace !== current_front_face) {
-                current_front_face = render.FrontFace;
-                game.Gl.frontFace(render.FrontFace);
-            }
-
-            switch (render.Kind) {
-                case RenderKind.Textured:
-                    draw_textured(game, transform, render);
-                    break;
-            }
-        }
-    }
-}
-
-function use_textured(game: Game, material: Material<TexturedLayout>) {
-    game.Gl.useProgram(material.Program);
-    game.Gl.uniformMatrix4fv(material.Locations.Pv, false, game.Camera!.Pv);
-}
-
-function draw_textured(game: Game, transform: Transform, render: RenderTextured) {
-    game.Gl.uniformMatrix4fv(render.Material.Locations.World, false, transform.World);
-
-    game.Gl.activeTexture(GL_TEXTURE0);
-    game.Gl.bindTexture(GL_TEXTURE_2D, render.Texture);
-    game.Gl.uniform1i(render.Material.Locations.Sampler, 0);
-
-    game.ExtVao.bindVertexArrayOES(render.Vao);
-    game.Gl.drawElements(render.Material.Mode, render.Mesh.IndexCount, GL_UNSIGNED_SHORT, 0);
-    game.ExtVao.bindVertexArrayOES(null);
-}
