+++ materials/layout_textured.ts
